Created a module Helper Enigma
