+++ spec/helper.rb
@@ -1,2 +1,3 @@
-require 'simplecov'
-SimpleCov.start
+# require 'simplecov'
+# SimpleCov.start
+# require 'rspec'
